delete items on tmpfs file system, batrash others
I don't often see a purpose for a trash can on a tmpfs file system, so I prefer to delete straight away on a tmpfs. But when I have to choose each time between trashing or removing to acomplish that, I might make the wrong choice when I am by accident working on a non-tmpfs. Therefor I hard coded this choice in a plugin for spacefm, my preferred file manager.